requirements.txt
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.2
tqdm==4.66.2